Terrain smoothing
It took me most of Sunday but I finally arrived at a terrain smoothing
solution I like.
It uses a lot of brute force, sub optimal logic, but it is still faster
that that random A* solution I'm using.
I can now decimate the height map and have the A* run on a fraction of
the original scope.
I can now reduce the number of bezier currve nodes to crate a smoother
set of turns in the road.
When it comes to smoothing the terrain, I am sampling the road at
regular intervals, generating a list of poins.
I will then go through the actual A* path nodes and collect all
surrounding terrain vertices (lots of duplicated vertices picked up, but
they end up in a hash set)
For each vertex I will then run through the path vertices and find the
closest one. This is the one that will contribute the correct height.
This is a brute force approach, but might be sufficient. This are of
code is already quite complex.

Next up:
-making the road transition smoothly to the next track section
-stitching up two generated track section
-threding the generation of the next section (maybe with A*
improvements)